swap eventdaterange and daterange back over
